rmend_project/requirements.txt
Django==3.1.2
psycopg2-binary==2.8.5
django-cors-headers==3.5.0
djangorestframework==3.12.2
djangorestframework-gis==0.16